+++ packages/zendesk-adapter/src/adapter.ts
@@ -74,7 +74,7 @@ import slaPolicyFilter from './filters/sla_policy'
 import macroAttachmentsFilter from './filters/macro_attachments'
 import omitInactiveFilter from './filters/omit_inactive'
 import tagsFilter from './filters/tag'
-import hcLocalesFilter from './filters/help_center_locale'
+import hcLocalesFilter from './filters/guide_locale'
 import webhookFilter from './filters/webhook'
 import targetFilter from './filters/target'
 import defaultDeployFilter from './filters/default_deploy'
@@ -84,23 +84,23 @@ import handleAppInstallationsFilter from './filters/handle_app_installations'
 import referencedIdFieldsFilter from './filters/referenced_id_fields'
 import brandLogoFilter from './filters/brand_logo'
 import articleFilter from './filters/article'
-import helpCenterFetchArticle from './filters/help_center_fetch_article'
+import helpCenterFetchArticle from './filters/guide_fetch_article'
 import articleBodyFilter from './filters/article_body'
 import { getConfigFromConfigChanges } from './config_change'
 import { dependencyChanger } from './dependency_changers'
 import customFieldOptionsFilter from './filters/add_restriction'
 import deployBrandedGuideTypesFilter from './filters/deploy_branded_guide_types'
 import { Credentials } from './auth'
-import hcSectionCategoryFilter from './filters/help_center_section_and_category'
-import hcTranslationFilter from './filters/help_center_translation'
-import fetchCategorySection from './filters/help_center_fetch_section_and_category'
-import hcParentSection, { addParentFields } from './filters/help_center_parent_to_section'
-import hcGuideSettings from './filters/help_center_guide_settings'
+import hcSectionCategoryFilter from './filters/guide_section_and_category'
+import hcTranslationFilter from './filters/guide_translation'
+import fetchCategorySection from './filters/guide_fetch_section_and_category'
+import hcParentSection, { addParentFields } from './filters/guide_parent_to_section'
+import hcGuideSettings from './filters/guide_guide_settings'
 import removeBrandLogoFilter from './filters/remove_brand_logo_field'
 import categoriesOrderFilter from './filters/guide_order/categories_order'
 import sectionsOrderFilter from './filters/guide_order/sections_order'
 import articlesOrderFilter from './filters/guide_order/articles_order'
-import hcServiceUrl from './filters/help_center_service_url'
+import hcServiceUrl from './filters/guide_service_url'
 import everyoneUserSegementFilter from './filters/everyone_user_segment'
 import hcLanguageSettings from './filters/guide_language_translations'
