+++ packages/zendesk-adapter/src/change_validators/guide_activation.ts
