+++ packages/zendesk-adapter/src/change_validators/index.ts
@@ -40,8 +40,8 @@ export { articleLabelNamesRemovalValidator } from './article_label_names_removal
 export { everyoneUserSegmentModificationValidator } from './everyone_user_segment_modification'
 export { brandFieldForBrandBasedElementsValidator } from './brand_field_for_branded_based_elements'
 export { translationForDefaultLocaleValidator } from './translation_for_default_locale'
-export { helpCenterActivationValidator } from './help_center_activation'
-export { helpCenterCreationOrRemovalValidator } from './help_center_creation_or_removal'
+export { helpCenterActivationValidator } from './guide_activation'
+export { helpCenterCreationOrRemovalValidator } from './guide_creation_or_removal'
 export { categoriesOrderValidator } from './guide_order/categories_order_validator'
 export { sectionsOrderValidator } from './guide_order/sections_order_validator'
 export { articlesOrderValidator } from './guide_order/articles_order_validator'
