+++ packages/zendesk-adapter/src/change_validators/translation_for_default_locale.ts
@@ -21,7 +21,7 @@ import {
 import Joi from 'joi'
 import { createSchemeGuardForInstance, resolveValues } from '@salto-io/adapter-utils'
 import { collections } from '@salto-io/lowerdash'
-import { isTranslation, TranslationType } from '../filters/help_center_section_and_category'
+import { isTranslation, TranslationType } from '../filters/guide_section_and_category'
 import { lookupFunc } from '../filters/field_references'
 import { ARTICLE_TYPE_NAME } from '../constants'
