+++ packages/zendesk-adapter/src/filters/guide_fetch_article.ts
