+++ packages/zendesk-adapter/src/filters/guide_locale.ts
