+++ packages/zendesk-adapter/src/filters/guide_service_url.ts
