+++ packages/zendesk-adapter/test/change_validators/guide_activation.test.ts
