+++ packages/zendesk-adapter/test/change_validators/guide_creation_or_removal.test.ts
