+++ packages/zendesk-adapter/test/filters/guide_section_and_category.test.ts
@@ -21,7 +21,7 @@ import {
   ObjectType, ReferenceExpression,
   toChange,
 } from '@salto-io/adapter-api'
-import filterCreator from '../../src/filters/help_center_section_and_category'
+import filterCreator from '../../src/filters/guide_section_and_category'
 
 
 import { ZENDESK } from '../../src/constants'
